Export room files in an array
This will make it easier when it comes to adding new rooms later in development. Any new rooms I add should be added in the index.js file for the scenes directory. Hopefully I can create a script to automate this process for me so I don't have to write in the entries myself.

The array will be read from later on for a changeRoom event listener to load new scenes on event dispatch.