Event listener structure created
Whenever you want to switch to another room, an event listener is fired containing the constructor for that room to the main index.js for the site. The site then uses the constructor to create a new currentScene.

The list of rooms was moved to home because constructors can be added as properties to the panel meshes so when they are clicked on the constructor can be accessed for event dispatching. All other rooms only need the link to the home room.